+++ config/set/php55.php
@@ -3,6 +3,8 @@
 declare(strict_types=1);
 
 use Rector\Php55\Rector\Class_\ClassConstantToSelfClassRector;
+use Rector\Php55\Rector\FuncCall\GetCalledClassToSelfClassRector;
+use Rector\Php55\Rector\FuncCall\GetCalledClassToStaticClassRector;
 use Rector\Php55\Rector\FuncCall\PregReplaceEModifierRector;
 use Rector\Php55\Rector\String_\StringClassNameToClassConstantRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
@@ -12,4 +14,6 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services->set(StringClassNameToClassConstantRector::class);
     $services->set(ClassConstantToSelfClassRector::class);
     $services->set(PregReplaceEModifierRector::class);
+    $services->set(GetCalledClassToSelfClassRector::class);
+    $services->set(GetCalledClassToStaticClassRector::class);
 };
