+++ rules-tests/Php55/Rector/FuncCall/GetCalledClassToSelfClassRector/Fixture/skip_trait.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php74\Rector\FuncCall\GetCalledClassToSelfClassRector\Fixture;
+namespace Rector\Tests\Php55\Rector\FuncCall\GetCalledClassToSelfClassRector\Fixture;
 
 trait SkipTrait
 {
