+++ rules-tests/Php55/Rector/FuncCall/GetCalledClassToStaticClassRector/Fixture/skip_anonymous_class.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php74\Rector\FuncCall\GetCalledClassToStaticClassRector\Fixture;
+namespace Rector\Tests\Php55\Rector\FuncCall\GetCalledClassToStaticClassRector\Fixture;
 
 new class {
     public function callOnMe()
