+++ rules-tests/Php55/Rector/FuncCall/GetCalledClassToStaticClassRector/Fixture/skip_on_final_class.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php74\Rector\FuncCall\GetCalledClassToStaticClassRector\Fixture;
+namespace Rector\Tests\Php55\Rector\FuncCall\GetCalledClassToStaticClassRector\Fixture;
 
 final class SkipOnFinalClass
 {
