+++ tests/Issues/GetCalledClassToSelfAndStatic/config/configured_rule.php
@@ -2,8 +2,8 @@
 
 declare(strict_types=1);
 
-use Rector\Php74\Rector\FuncCall\GetCalledClassToSelfClassRector;
-use Rector\Php74\Rector\FuncCall\GetCalledClassToStaticClassRector;
+use Rector\Php55\Rector\FuncCall\GetCalledClassToSelfClassRector;
+use Rector\Php55\Rector\FuncCall\GetCalledClassToStaticClassRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
